Move test and improve assertion
